refactor: move Snake and Food classes to separate files
